Fixed some bugs - OutUp v0.0.8

+++ controllers/training.js
@@ -1,7 +1,9 @@
 const Training = require('../models/Training');
+const Base = require('../models/Base');
 const errorHandler = require('../utils/errorHandler');
 const generate = require('../utils/generator');
 const serverPath = require('../localPath');
+const compressArray = require('../utils/compressArray');
 
 const fs = require('fs');
 const path = require('path');
@@ -20,6 +22,24 @@ module.exports.create = async function (req, res) {
             image: req.file ? req.file.path : '',
             key: generate()
         }).save();
+
+        let arrayExercises = [];
+        for (let days of training.weeks) {
+            for (let exercises of days['days']) {
+                for (let exercise of exercises['exercises']) {
+                    arrayExercises.push(String(exercise['exercise']));
+                }
+            }
+        }
+
+        const result = compressArray(arrayExercises);
+        for( let exercise of result){
+            await Base.findByIdAndUpdate(
+                { _id: exercise.value },
+                { $inc: { use: exercise.count } },
+                { new: true }
+            );
+        }
 
         res.status(201).json(training);
     } catch (e) {
@@ -51,15 +71,44 @@ module.exports.delete = async function (req, res) {
     try {
         const workout = await Training.findOne({ _id: req.params.id });
 
-        fs.unlink(path.resolve(serverPath.path, workout.image), async (err) => {
-            if (err) {
-                return res.status(500).json({ message: 'Failed to delete workout!' });
-            }
-            else {
-                await Training.findByIdAndDelete({ _id: req.params.id });
-                res.status(200).json({ message: 'Workout deleted!' });
-            }
-        });
+        if (workout.owner == req.user.id) {
+            fs.unlink(path.resolve(serverPath.path, workout.image), async (err) => {
+                if (err) {
+                    return res.status(500).json({ message: 'Failed to delete workout!' });
+                }
+                else {
+                    await Training.findByIdAndDelete({ _id: req.params.id });
+                    
+                    let arrayExercises = [];
+                    for (let days of workout.weeks) {
+                        for (let exercises of days['days']) {
+                            for (let exercise of exercises['exercises']) {
+                                arrayExercises.push(String(exercise['exercise']));
+                            }
+                        }
+                    }
+            
+                    const result = compressArray(arrayExercises);
+                    for( let exercise of result){
+                        await Base.findByIdAndUpdate(
+                            { _id: exercise.value },
+                            { $inc: { use: -exercise.count } },
+                            { new: true }
+                        );
+                    }
+
+                    return res.status(200).json({ message: 'Workout deleted!' });
+                }
+            });
+        }
+        else {
+            await Training.findByIdAndUpdate(
+                { _id: req.params.id },
+                { $pull: { users: req.user.id } },
+                { new: true }
+            );
+            res.status(200).json({ message: 'Workout deleted!' });
+        }
     } catch (e) {
         errorHandler(res, e);
     }
@@ -71,9 +120,8 @@ module.exports.update = async function (req, res) {
             { _id: req.body._id }
         );
 
-        let weeks = training.weeks;
         let i = 0, str = '';
-        for (let days of weeks) {
+        for (let days of training.weeks) {
             let j = 0;
             for (let exercises of days['days']) {
                 let t = 0;
@@ -88,7 +136,7 @@ module.exports.update = async function (req, res) {
             i++;
         }
 
-        const workout = await Training.findByIdAndUpdate(
+        await Training.findByIdAndUpdate(
             { _id: req.body._id },
             { $set: { [str]: req.body.status } },
             { new: true }
